+++ examples/P1P2-bridge-esp8266/P1P2_NetworkParams.h
@@ -2,7 +2,10 @@
  *
  * Copyright (c) 2019-2022 Arnold Niessen, arnold.niessen -at- gmail-dot-com  - licensed under GPL v2.0 (see LICENSE)
  *
+ * Version history
+ * 20220808 v0.9.15 extended verbosity command, unique OTA hostname, minor fixes
  * 20220802 v0.9.14 AVRISP, wifimanager, mqtt settings, EEPROM, telnet, outputMode, outputFilter, ...
+ * ..
  *
  * WARNING: P1P2-bridge-esp8266 is end-of-life, and will be replaced by P1P2MQTT
  *
@@ -26,7 +29,8 @@ char MQTT_CLIENTNAME[9] = "P1P2_xxx"; // Must be unique, include 4th byte of IPv
 #define MQTT_CLIENTNAME_IP 5          // start of xxx for ip nr in MQTT_CLIENTNAME
 
 // OTA
-#define OTA_HOSTNAME "P1P2MQTT"
-//#define OTA_PASSWORD "P1P2MQTT"
+char OTA_HOSTNAME[13] = "P1P2MQTT_xxx";
+#define OTA_HOSTNAME_PREFIXIP 9        // start of xxx for ip nr in OTA_HOSTNAME
+#define OTA_PASSWORD "P1P2MQTT"
 
 #endif /* P1P2_NetworkParams */