+++ examples/P1P2Monitor/P1P2Config.h
@@ -2,6 +2,7 @@
  *
  * Copyright (c) 2019-2022 Arnold Niessen, arnold.niessen -at- gmail-dot-com  - licensed under GPL v2.0 (see LICENSE)
  *
+ * 20220808 v0.9.15 Added compile date/time to verbosity report
  * 20220802 v0.9.14 New parameter-write method 'e' (param write packet type 35-3D), error and write throttling, verbosity mode 2, EEPROM state, MCUSR reporting,
  *                  pseudo-packets, error counters, ...
  *                  Simplified code by removing packet interpretation, json, MQTT, UDP (all off-loaded to P1P2MQTT)
@@ -27,6 +28,8 @@
 
 //#define SERIALSPEED 115200
 #define SERIALSPEED 250000
+
+#define WELCOMESTRING "* P1P2Monitor-v0.9.15"
 
 #define INIT_VERBOSE 3
 // Set verbosity level